Rake tasks and models to support scraping content
Turns out this is pretty simple to do as the NCS site has a sitemap pointing to the relevant content, and within both category and job profile pages the markup is nicely structured to allow parsing interesting content.

For now this only parses the list of job skills as a POC. We can add more as we find it necessary. It's also limited to only scraping 10 job profiles for now to avoid unnecessary load on NCS.