Introduce Version model
The intent here is to allow a future/draft version of the job profile and associated models to be created without breaking the existing job profile data in the meantime. Once the screen scraping completes successfully the draft version becomes the current version. I've also added support for keeping a previous version - not sure if that will be useful but no harm in adding it for now and removing if it proves unnecessary.

This means that when working with the core `JobProfile`, `Category` and `Skill` models we will instead need to refer to `Version.current.job_profiles`, `Version.current.categories` and `Version.current.skills` respectively.