GET-177 Add full text search for Job Profiles
jira-ticket: https://dfedigital.atlassian.net/browse/GET-177

When searching for Job Profiles on the check your skills page as
well as careers page, allow user to do a full text search for words
in Title, OR  alternative titles OR description. Add a ts_vector
index on the fields to speed up search.

To allow words to be used as prefixes use the ":*" notation to
be appended to each word's ending, and an OR between multiple words
to allow the query to search across words entered. Make sure to
clean up strings from obscure unicode characters and non alpha numeric
characters to maintain the correct syntax for SQL queries as well as
stop any injections, found in the Query search string text formatter.

Order query results by ts_vector name score returned. To avoid
injections as we are directly using a user entered string in the
query, quote the entry before using Arel.sql to ensure no injection
or malicious attempts make it through to the query.