Remove Turbolinks Turbolinks is intended to improve page load times, by avoiding the need to fetch and parse CSS and JS assets described in the `head` section of the page. In practice this works to a degree, but has complications.
Already we have had to adapt the progressive enhancement for search to listen to custom turbolinks events. The Azure App Insights tracking script is working intermittently since this is only firing on initial page load or manual refresh, not when user navigates to a new page by clicking links.

There are numerous anecdotal reports of various issues with the library, including double loading of pages, accessibility limitations etc. Overall, it's not adding value so good to remove.