GET-281 Imrpove CI build time
jira-ticket: https://dfedigital.atlassian.net/browse/GET-281

* Download RVM to use our own ruby version 2.6.3
* Add path to other tasks with: [task.prependpath] as
profile/bashrc loaded in other tasks reads from readonly
system profile/bashrc
* Move tests out of docker build to run on box
* Add caching for bundle step. Unfortunately the caching
step does not return the bin files as executables
as loses the symlinks. So we have a step to allow
files modes to be changed. Also note that bundle always
runs as (there is no condition to say it doesn't run with
caching is restored), this is to restore the correct symlinks. Bundle
running again does not cause any slowness if caching is restored
so it is ok until caching is fixed.