GET-363 Separate different job profiles in skill builder
jira-ticket: https://dfedigital.atlassian.net/browse/GET-363

When a user navigates to a job profile and selects a list of skills,
the scope of those skills are only to the job profile. If a user navigates
back to the search results and picks a different job profile, they should see
all skills ticked unless they have visited that page before. Moving towards
a skills scoped to a job profile we make it easier to move to the skill builder
v2 which will now only require FE changes only.