GET-232 Add Skill Matcher
jira-ticket: https://dfedigital.atlassian.net/browse/GET-232

Add skill matcher, which accepts a session with a set of skills, and optionally
a job profile id to scope to.

If there is no job profile id, the skills matcher finds all job profile ids
matching those skills, and groups them by skills matched count of job profiles as a subquery.
Then by grouping on the count of skills matched we get all job profile ids that have the same
score, and order them in ascending order of name, as well as on frequency. Finally we select
the records scoped to the fields needed and expand the ids of the ordered group maintaining
ordinality.

For the score we create a new hash from the ids and calculate the score mapped to a job profile id.
The scoring system chosen counts the skills matched in the job profile / the total count of skills selected %