Address the following edge cases:
1. When a user has reached the capacity of max 5 profiles
when trying to access Check your existing skills one will
get redirected to Your Skills page as one should not be able
to search anymore.

2. When a user has a set of profiles already persisted on
the session, and decides to edit one, unticks all the skills
and goes back to the Your Skills page, one should not see that
profile in there anymore.