GET-445 Change session data column to jsonb
jira-ticket: https://dfedigital.atlassian.net/browse/GET-445

Reading data from a jsonb column orders hash key values. Since
we wish to keep the order of the job profile ids, add a new key
that maps to an array of ids, and read from that instead. This way
we insure the order of job ids is maintained

Setup heroku review apps deployment. We use DATABASE_URL for both dbs if
app is deployed to heroku