GET-450 Add user and passwordless migrations and models
jira-ticket: https://dfedigital.atlassian.net/browse/GET-450

Add passwordless gem to handle authentication and magic link.
Passwordless requires a passwordless sessions table to store tokens.
We also allow the tokens to only be used once by setting the default
value to true and adding the claimed_at column to the table.

Passwordless allows emails to be sent by actionmailer. We will be using
Notify to add placeholder till we tackle that next.

In terms of session expiration, I am going with adhering with the defaults
for now (1.year). We can override if we aren't happy with that.

As for sign in tokens expiration I am also going with the default fo 1 hour unless
we aren't happy with that either.

To use passwordless, generate a user model with an email that should be unique.
We are using citext to not worry about emails being case sensitive.
A user model will be associated with one session at a time for now.