Turn off HTML5 validation
number_field turns the HTML5 validation on by default,
so we need to make sure it's off.

Also this field will allow user to type some of the following
non-numeric characters:

A floating-point number consists of the following parts,
in exactly the following order:

Optionally, the first character may be a "-" character.
One or more characters in the range "0—9".
Optionally, the following parts, in exactly the following order:
a "." character
one or more characters in the range "0—9"
Optionally, the following parts, in exactly the following order:
a "e" character or "E" character
optionally, a "-" character or "+" character
One or more characters in the range "0—9".

https://stackoverflow.com/questions/31706611/why-does-the-html-input-with-type-number-allow-the-letter-e-to-be-entered-in

This PR also fixes a small typo.