GET-412 Fix bug in sign in and merging sessions
jira-ticket: https://dfedigital.atlassian.net/browse/GET-27

When a user signs in, their session is authenticated in the passwordless gem,
outside the user_session model. This means that it does not have the version
instantiated and added to the session. When the session does go through the user
session model, it is seen as an obselete session and is destroyed due to mismatch in
the version.

To fix this, call the session directly in the callback after signing in, to not
risk destroying the session, and merging old data correclty without losing the
authentication/