GET-412 Retain email on validation on sign in, and add anchor to url
When a user adds an invalid email and is redirected back from the sign in
page, we should retain the email. We pass this as a query param in the redirect url.
Since `redirect_back` does not allow us to do this, build this functionality ourselves.
Also add an anchor to the email to make sure we scroll back down to the sign in form.

Move previous redirect logic to same URL Parser class as it shares the url is part of host
functionatliy.