Skip PID/Postcode Eligibility pages
Skip PID/Postcode eligibility pages if the the information
has already been captured.

When user clicks start now:

a) if new session, go to postcode (eligibility) page
b) if user has a postcode (eligibility) but no personal
data in session, redirect to personal data page
c) if user has a postcode and personal data in session,
redirect to tasklist

When saving/returning to the journey, the same behaviour applies
for a user that has previously supplied the data.