Explicitly order by ordinality number on results
When unnesting an aggregate array with ordinality with minimal syntax:
`WITH ORDINALITY ordered_id`, this was minimising the following
syntax:
`WITH ORDINALITY some_table(ids, ordinality_number)`
But this does not guarantee the order of the unnested arrays is maintained.
Explicitly ordering on the ordinality number as a secondary order
after the skills match makes sure the order is retained.
`