Header logo accessibile
Voice over detects a missing image description. This is an
SVG(graphic) and according to this example:

https://developer.paciellogroup.com/blog/2013/12/using-aria-enhance-svg-accessibility/

It needs the Title and Description to provide context on the image.

Also, according to:

https://design-system.service.gov.uk/styles/images/

The alt on the img nested under the SVG does not need a
non-empty alt as the context has been already provided upper.