GET-538 Do not destroy session if versions are different
jira-ticket: https://dfedigital.atlassian.net/browse/GET-538

When users were hitting the page for the first time we were deleting
the session from the database as they didn't have a version in the data.
This was causing unnecessary read/write to the database also causing the
database to seem like it was "skipping" ids, which were records created and
deleted unnecessarily.

This also caused an issue with authenticated users linked to sessions. If a user was signed
in with a session and the version changed, which caused their session to be deleted they could
not log in again, but rather see an error where we were trying to access a non existing session.

Now we "clear" the session instead of deleting it, only if the session has data, meaning there are no
unnecessary writes to the db, and we don't risk deleting a session linked to a user.

However I am unhappy with this versioning solution, as after we persist sessions this
fix seems more of a pain than a solution. We now have control over sessions, and can make
sure to be backward compatible when releasing different versions. I propose after going live
with phase 3 we get rid of versioning, as it is needed for this release only. I added a commented
out test that shows another failure this causes and as a reminder for us to remove after the release.