Remove AuthenticationHelper
This helper is no longer needed as the same
functionality can be achieved with this change.

The current flow will redirect to /return-to-saved-results
after user submits the form. However, due to the redirect we
are losing the object and its validation errors (if any).

We also have to make sure that when deeplinking into this page
we do not trigger a validation for the user.

To tackle these issues, we only validate the object if the
email key is on the params hash, as that means the form has
been submitted and the user is currently redirected to
/return-to-saved-results.