GET-605 Add Job Vacancy search
jira-ticket: https://dfedigital.atlassian.net/browse/GET-605

Job vacancy search class takes in options for the API search
and validates the postcode before querying teh API.

We have two methods jobs and count. We memoize the API query so we
only do it once for either. This can be reused for the job profile
page as well as the job vacancy search. We wrap Job vacancies with
and object to extract exact fields

Since pagination in the API is to 50 jobs vs ours which is 10, convert
the pages before passing the mapping we want into the API