Fix IE11 compatibility issue
We currently have for event tracking that is using
forEach function. However, IE11 throws the following
error:

Object doesn't support property or method 'forEach'

Because of this the behaviour our accordion component
is broken on IE11 giving bad user experience.

Decided to go with the simplest solution of using the
classic for loop.