All outbound links open in the same tab
Now that we have these links tracked in GA, we can drop
the tracked-event data attribute as we don't need to track
them in AppInsights.

NOTE:

We can not yet get rid of TrackEvents() JS function as
we still need to track the interraction on our inpage
smart survey. Once we implement GA backend tracking,
we can drop this script.