Cache all static assets
We set static cachecontrol to 'public, max-age=31536000'
so the browser caches all static assets for up to a year.
In production every asset has a hash added to its name,
so whenever the file changes the browser requests
the latest version as the hash and therefore the whole
filename changes.

Useful resource:

https://blog.codeship.com/use-a-cdn-for-all-static-assets/