Introduce IT training questions page
We are treating this as a separate which
could future have additional options for IT training
as per conversation with Shelina.

This approach is also less complexity than having to
use the same :training key on the session for both
it training & training pages.

The sequence is:

Maths or English skills
IT Training - Computer skills
Job Hunting skills

https://dfedigital.atlassian.net/browse/GET-831