Update Job Profile page markup
Following changes are introduced:

1. Select type of work changes into Apply to become section
2. This new section since it's the last and has a special style
for its delimiter, we are moving the dynamic sections' delimiters
on top of each section rather than after, so that the last dynamic
section's delimiter does not collide with the custom one that belongs
to Apply to become section.
3. Since concat/+ in ruby escape strings, we need to call an html_safe
on the our dynamic sections' bodies when prepending the delimiter.