GET-854 Add role button to links styled as buttons which affect journey
jira-ticket: https://dfedigital.atlassian.net/browse/GET-854

Add role button to all buttons which affect a user's journey. As per
recommendation from DAC report:
If continuing through the service it is recommended that ‘Continue’ buttons are used for consistency purposes and to inform the user that they’ll be continuing throughout the service.
The ‘button’ is still a link in semantics, and has not been given a role of button as an alternative
page 41