Filter out dob from Active Admin forms on PID
We do not wish to leak the date of birth data in our logs
as it's sensitive information.
Adding dob to the rest of the filtered params we already have
will fix the issue.

Active admin has its own convention when dealing with date types.

In our case it will use dob(1i), dob(2i), dob(3i).