GET-1020 Add importer for course lookup
jira-ticket: https://dfedigital.atlassian.net/browse/GET-1020

Generate course lookups for opportunities. These should fall under
the relevant subjects of ESOL, English, Maths. If an opportunity does
not fall under these subjects do not create it. Also a lookup either
belongs to a venue if it exists or a provider.