Add tabs to skills matcher page
Also add tabs to job profiles search page. The way
the original GDS component works needs a bit of tweaking
as follows:

The inactive tab has the link to the destination page and
not an anchor link as the component would expect. To achieve
that the inactive tab must not have the govuk-tabs__tab class,
otheriwise it will error as it's expecting something like:
'#element-id' which would not work well in our case.