Add the Back link to the following sections:
1. Save your progress journey
2. Return to your saved progress journey
3. Training questions

Because of a few edge cases, the back_link is now able
to accept a custom url for redirection and can also ignore
paths.

The current Back link implementation needs to ignore
paths like:

- email-sent-again which is only accessible via POST
- link-sent again which is only accessible via POST

Training questions, because they come in sequence, and
according to GDS pattern the user needs to be able to navigate
throuhg every question using the Back link, so adding
explicit redirect routes fixes this issue.