GET-1042 Add sector and hierarchy to search ranking
jira-ticket: https://dfedigital.atlassian.net/browse/GET-1042

Amend scores of searching as the sum was being outputed before.
Add sector to partial search to boost profiles that include
a sector match. Hierarchy however needs to be a secondary rank
which we order in ascending order to push down. If hierarchy was
added to normal order it skews the results.