Authentication process to ignore the resource name
We recently noticed that resources in Azure Active Directory
can have the name attribute changed breaking our current
implementation which relied on the combination of
resource_id + email + name as the unique identifier
of authorised users for the Admin feature.

Relying instead on just resource_id + email as the unique
identifier will fix this issue.