Remove the IP Restriction config
Initially we tried adding 0.0.0.0/0 as an additional
entry to the existing configuration. This means that
any IP is allowed access. However, there seems to still
be an access issue for people with the DfE KIT on.

Because we don't have the DfE KIT on it's hard to debug
what the issue is. To unblock things, we decided to remove
the IP Restriction config completely.