Hotfix: Overcome webdriver connection issue
Running the specs in offline mode, seem to throw
in some cases the following error:

Webdrivers::ConnectionError:
  Can not reach https://chromedriver.storage.googleapis.com/

This is coming from spec/support/personal_data when trying
to deep-link straight into the PID form and it looks
like an interaction issue with Webdriver.

Getting to PID from the landing page seems to solve the issue,
as the cookies modal is accepted at that level now.

Also, removes the need to identify the specs that have the :js
tag.