Do not allow methods outside accepted HTTP methods
Frontdoor has a security check routine that is ran
every couple of hours. Among other checks, it tries
to call a DEBUG method against our server.

As a good security practice the server should respond
with method not allowed in this case. This will also
turn off the noise in Sentry.