Override target job when alternative job is passed
We want to allow users to surface additional job vacancies,
by being able to override the targetted job on the session.

Also fix the following bug:

until now, if a page/distance param was not being passed
at the controller level, it was still explicitly passed as nil in
the JobVacancySearch causing the initial page
number to be nil.to_i = 0 instead of 1.
similar situation for distance.