Add skills_criss_cross db view
In essence, the view translates any skill that is associated with a
master skill to either that master skill, or to any other skill
associated with the master skill.
It works in both directions, so you don't have to second-guess whether
a skill is a master or a variant (or neither); it also functions
regardless of whether any records have the skill master name
set in skills table.